docs: record vue.html S2 DOM installment
