+++ crates/vize_ricalco/src/emit.rs
@@ -64,6 +64,7 @@ mod flag;
 mod fragment;
 mod helper;
 mod hoist;
+mod html;
 pub(crate) mod js;
 mod memo;
 mod merge;
