+++ crates/vize_ricalco/src/emit/outlet.rs
@@ -1,9 +1,7 @@
 //! `<slot>` outlets: `_renderSlot(_ctx.$slots, name, props?, fallback?)`.
 //!
 //! Matches `vize_atelier_core::codegen/slots/outlet.rs` plus the if/for
-//! branch shapes. Static prop keys are camelized. A component that
-//! forwards an outlet uses `_: 3 /* FORWARDED */` unless it sits inside
-//! a scoped `withCtx` (then `_: 2` + `DYNAMIC_SLOTS`).
+//! branch shapes. Static prop keys are camelized.
 
 use vize_s0::camelize;
 use vize_s2::expr::{ExprRef, OpaqueReason};
@@ -169,11 +167,7 @@ fn emit_props(cx: &mut EmitCx<'_>, slot: &SlotOp<'_>, key: Option<&str>) -> Resu
                 }
             }
             Piece::VueHtml(html) => {
-                cx.buf.push("innerHTML: ");
-                match super::props_object::html_value(html)? {
-                    Some(source) => cx.buf.push(source),
-                    None => cx.buf.push("undefined"),
-                }
+                super::html::emit_pair(cx, html)?;
             }
             Piece::On(_)
             | Piece::ModelValue { .. }
