+++ davinci-road/plan/README.md
@@ -56,7 +56,7 @@ and dependency contract shared by every phase.
 | ---------------------------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
 | [phase-0.md](./phase-0.md)                                                                                       | Instrumentation and groundwork                                                                  | **Executed 2026-08-14** — all 13 tasks landed; reference-runner recordings pending (see exit gate)                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        |
 | [phase-1.md](./phase-1.md)                                                                                       | One arena, real expressions                                                                     | **Executed 2026-08-14 → 2026-08-15, exit gate not clean** — 11 of 13 tasks landed at the original exit, corpus byte-parity held; P1-8 has since been resolved by the scanner deletion follow-up, P1-12 awaits sign-off; see the exit gate for remaining blockers                                                                                                                                                                                                                                                                                                                                                                          |
-| [phase-2.md](./phase-2.md) + [phase-2-tasks.md](./phase-2-tasks.md) + [phase-2-records.md](./phase-2-records.md) | Disegno and the pass manager                                                                    | **Executing since 2026-08-19** — re-cut 2026-08-17 at the P1 exit with all 20 IDs expanded to the full contract (P2-5 and P2-12 split into a/b, 22 tasks). **Current ledger (2026-08-28): 16 of 22 complete; 2 active and blocked (P2-9, P2-11); 4 untouched and dependency-blocked (P2-12b, P2-16, P2-17, P2-20).** P2-11 has 32 landed installments through [#5198](https://github.com/ubugeeei-prod/vize/pull/5198). Corpus scope is the executable inventory: 146 gitlinks / 142 ecosystem projects. The phase file owns the current ledger and exit gate; the records index separates current evidence from historical task records. |
+| [phase-2.md](./phase-2.md) + [phase-2-tasks.md](./phase-2-tasks.md) + [phase-2-records.md](./phase-2-records.md) | Disegno and the pass manager                                                                    | **Executing since 2026-08-19** — re-cut 2026-08-17 at the P1 exit with all 20 IDs expanded to the full contract (P2-5 and P2-12 split into a/b, 22 tasks). **Current ledger (2026-08-29): 16 of 22 complete; 2 active and blocked (P2-9, P2-11); 4 untouched and dependency-blocked (P2-12b, P2-16, P2-17, P2-20).** P2-11 has 33 landed installments through [#5200](https://github.com/ubugeeei-prod/vize/pull/5200). Corpus scope is the executable inventory: 146 gitlinks / 142 ecosystem projects. The phase file owns the current ledger and exit gate; the records index separates current evidence from historical task records. |
 | [phase-3.md](./phase-3.md)                                                                                       | Impeto and backend convergence                                                                  | Drafted, provisional — re-cut at P2 exit                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  |
 | [phase-4.md](./phase-4.md)                                                                                       | Consumer convergence                                                                            | Drafted, provisional — re-cut at P3 exit                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  |
 | [phase-5.md](./phase-5.md)                                                                                       | Incrementality substrate                                                                        | Drafted, provisional — re-cut at P4 exit                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  |
