+++ davinci-road/plan/croquis-consumption.md
@@ -379,7 +379,7 @@ The naive lane counts raw word-boundary text matches per product name (`\.field`
 | `ScopeId`                              |       37 |   56 | `vize_canon` (13/18), `vize_croquis_cf` (23/33), `vize_maestro` (1/2), `vize_patina` (0/1), `vize_vitrine` (0/2)                                                                                                                                                                                                                                                                                   |
 | `ScopeKind`                            |      105 |  127 | `vize_atelier_core` (1/2), `vize_canon` (36/47), `vize_croquis_cf` (21/25), `vize_maestro` (43/49)                                                                                                                                                                                                                                                                                                 |
 | `SlotUsage`                            |        7 |   17 | `vize_canon` (1/3), `vize_croquis_cf` (3/8), `vize_maestro` (3/6)                                                                                                                                                                                                                                                                                                                                  |
-| `Span`                                 |        0 |  569 | `vize_atelier_core` (0/2), `vize_atelier_jsx` (0/63), `vize_atelier_sfc` (0/6), `vize_atelier_vapor` (0/1), `vize_canon` (0/56), `vize_carton` (0/89), `vize_davinci` (0/20), `vize_maestro` (0/4), `vize_patina` (0/145), `vize_relief` (0/8), `vize_ricalco` (0/66), `vize_s2` (0/109)                                                                                                           |
+| `Span`                                 |        0 |  571 | `vize_atelier_core` (0/2), `vize_atelier_jsx` (0/63), `vize_atelier_sfc` (0/6), `vize_atelier_vapor` (0/1), `vize_canon` (0/56), `vize_carton` (0/89), `vize_davinci` (0/20), `vize_maestro` (0/4), `vize_patina` (0/145), `vize_relief` (0/8), `vize_ricalco` (0/66), `vize_s2` (0/111)                                                                                                           |
 | `SpreadProp`                           |        5 |    9 | `vize_canon` (5/9)                                                                                                                                                                                                                                                                                                                                                                                 |
 | `Symbol`                               |        0 |   68 | `vize_atelier_core` (0/5), `vize_atelier_sfc` (0/4), `vize_atelier_ssr` (0/2), `vize_canon` (0/3), `vize_croquis_cf` (0/27), `vize_fresco` (0/2), `vize_maestro` (0/1), `vize_patina` (0/15), `vize_relief` (0/5), `vize_ricalco` (0/1), `vize_vitrine` (0/3)                                                                                                                                      |
 | `SymbolFlags`                          |        0 |    5 | `vize_atelier_sfc` (0/5)                                                                                                                                                                                                                                                                                                                                                                           |
