+++ davinci-road/plan/phase-2-records/p2-11.md
@@ -14,7 +14,7 @@ atelier_dom **test space** (dev-dep carve-out). The old DOM lane is
 still the shipped compile path. The exit gate is unchanged.
 
 The `emit.rs` module docs keep a running capability list. **This
-record's landed numbers are the thirty-two `origin/main` PRs**, not that
+record's landed numbers are the thirty-three `origin/main` PRs**, not that
 comment.
 
 ## Installment table
@@ -53,6 +53,7 @@ comment.
 | 30  | 2026-08-28 | [#5178](https://github.com/ubugeeei-prod/vize/pull/5178) | `2299a2114` | elide inert slot-template bindings           |
 | 31  | 2026-08-28 | [#5183](https://github.com/ubugeeei-prod/vize/pull/5183) | `f5aa60553` | emit inline slot-template carriers           |
 | 32  | 2026-08-28 | [#5198](https://github.com/ubugeeei-prod/vize/pull/5198) | `2be66b0f0` | emit `v-show` runtime directives             |
+| 33  | 2026-08-29 | [#5200](https://github.com/ubugeeei-prod/vize/pull/5200) | `13cff4d99` | emit `v-html` raw HTML props                 |
 
 `landed` is the UTC committer date on `origin/main` (the merge-to-main
 clock). Per-installment records live in [p2-11/](./p2-11/):
@@ -89,8 +90,9 @@ clock). Per-installment records live in [p2-11/](./p2-11/):
 - [Installment 30 — inert slot-template bindings](./p2-11/installment-30.md)
 - [Installment 31 — inline slot-template carriers](./p2-11/installment-31.md)
 - [Installment 32 — `v-show` runtime directives](./p2-11/installment-32.md)
+- [Installment 33 — `v-html` raw HTML props](./p2-11/installment-33.md)
 
-## Current named remainder (after #5198)
+## Current named remainder (after #5200)
 
 Installment 20 closes the dynamic-argument `v-bind` key/modifier item that the
 installment-19 snapshot correctly left open. The current `emit.rs` capability
@@ -137,6 +139,12 @@ admits `vue.show`, Folio round-trips `vue.show value=`, and the DOM emitter
 wraps native elements and components in runtime `vShow` directive entries while
 preserving shipped helper order, patch flags, `v-for` empty-props shape, custom
 directive order, and native `v-model` listener keys.
+Installment 33 realizes `v-html` directly from S2: the lowering admits
+`vue.html` on elements and slot outlets, Folio round-trips value-bearing and
+value-less forms, and the DOM emitter writes shipped-compatible `innerHTML`
+props with exact patch-key order across native, component, slot, structural,
+object-spread, custom directive, and native `v-model` cases. Argument and
+modifier spellings remain deferred.
 
 The series is not done. Waiver budget remains zero. The old DOM lane remains
 the shipped path; the publish-graph decision, hydrated full-corpus comparison
