+++ davinci-road/plan/storage-inventory.tsv
@@ -43,12 +43,12 @@ s1_to_s2	emit	crates/vize_ricalco/src/emit/directive.rs	1	6	0	0	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/emit/filter.rs	0	0	1	1	0	0	0	0
 s1_to_s2	emit	crates/vize_ricalco/src/emit/hoist.rs	1	4	1	7	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/emit/js.rs	0	0	1	10	0	0	0	0
-s1_to_s2	emit	crates/vize_ricalco/src/emit/merge.rs	1	8	1	1	0	0	0	0
+s1_to_s2	emit	crates/vize_ricalco/src/emit/merge.rs	1	8	1	2	0	0	0	0
 s1_to_s2	emit	crates/vize_ricalco/src/emit/model.rs	1	6	1	4	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/emit/memo.rs	0	0	1	0	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/emit/model_key.rs	0	0	1	7	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/emit/on.rs	0	0	1	5	0	0	1	3
-s1_to_s2	emit	crates/vize_ricalco/src/emit/props.rs	1	2	1	2	0	0	0	0
+s1_to_s2	emit	crates/vize_ricalco/src/emit/props.rs	1	2	1	3	0	0	0	0
 s1_to_s2	emit	crates/vize_ricalco/src/emit/props_bind.rs	1	5	1	2	0	0	0	0
 s1_to_s2	emit	crates/vize_ricalco/src/emit/props_object.rs	1	5	0	0	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/emit/props_object_merge.rs	0	0	1	1	0	0	0	0
@@ -66,6 +66,7 @@ s1_to_s2	lower	crates/vize_ricalco/src/lower/directive.rs	1	6	0	0	0	0	0	0
 s1_to_s2	lower	crates/vize_ricalco/src/lower/element.rs	1	2	0	0	1	4	0	0
 s1_to_s2	-	crates/vize_ricalco/src/lower/expr.rs	0	0	1	3	0	0	0	0
 s1_to_s2	lower	crates/vize_ricalco/src/lower/forop.rs	1	2	1	5	1	2	0	0
+s1_to_s2	-	crates/vize_ricalco/src/lower/html.rs	0	0	1	2	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/lower/leaf.rs	0	0	1	5	1	4	0	0
 s1_to_s2	-	crates/vize_ricalco/src/lower/once_memo.rs	0	0	1	3	0	0	0	0
 s1_to_s2	-	crates/vize_ricalco/src/lower/show.rs	0	0	1	2	0	0	0	0
