+++ davinci-road/roadmap.md
@@ -113,10 +113,10 @@ holds or improves — this phase should be a measurable win, not a wash.
   timing, remark emission) land with the pass manager itself, so the DevTool's
   data feed exists from the first S2 build.
 
-**Current execution ledger (2026-08-28):** [16 of 22 tasks are complete](./plan/phase-2.md#current-execution-ledger-2026-08-28).
+**Current execution ledger (2026-08-29):** [16 of 22 tasks are complete](./plan/phase-2.md#current-execution-ledger-2026-08-29).
 P2-9 and P2-11 are active blocked series; P2-12b, P2-16, P2-17 and P2-20
-are untouched because their dependency gates remain open. P2-11 has 32 landed
-installments through [#5198](https://github.com/ubugeeei-prod/vize/pull/5198).
+are untouched because their dependency gates remain open. P2-11 has 33 landed
+installments through [#5200](https://github.com/ubugeeei-prod/vize/pull/5200).
 The executable fixture inventory is 146 gitlinks / 142 ecosystem projects;
 fixture checkout hydration is deliberately not a project-count source.
